requirements.txt
Django==1.8.5
Markdown==2.6.2
Pygments==2.0.2
Routes==2.2
django-filter==0.11.0
django-guardian==1.3
djangorestframework==3.2.4
psycopg2==2.6.1
pymongo==3.0.3
repoze.lru==0.6
six==1.10.0
wsgiref==0.1.2
